+++ packages/lemon_js_ui/lib/src/renderer/quickjs_ui_media_file_io.dart
@@ -65,7 +65,7 @@ Widget buildQuickjsUiFileSvg(
 }
 
 final class _QuickjsUiSvgFileLoader extends SvgFileLoader {
-  const _QuickjsUiSvgFileLoader(File file) : super(file);
+  const _QuickjsUiSvgFileLoader(super.file);
 
   @override
   String provideSvg(void message) =>